requirements.txt
Django
django-extensions
whitenoise
Pillow
django-ckeditor
django-resized
